Support query by user relation
A new predicate called functional predicate has been created just for this
kind of query. When query with a relation, a join has to been defined
with a new alias for the joining table (the table on the right side of
`LEFT JOIN`).

Functional predicate can be combined with other predicate using
compound predicate (And, Or, Not). Functional predicate supports
friend and follow relations plus inward, outward and mutual direction.

refs #356