Pass the configuration of skygear into plugin
The configuration is moved to skyconfig package. Now the main.go will pass the
configuration into plugin Context. And individual plugin will be able to query
the config object as needed.

In this commit, the exec transport will read the db config and pass it to exec
by os env.

refs #538