Fix ACL in query without user
When there is no user for the query, the ACL is ignored because
ReadableBy is empty. The builder should only ignore ACL if the
database is a private database.

ReadableBy is renamed as ViewAsUser to avoid the variable semantics
to be overloadded with whether to apply ACL. If ViewAsUser is nil,
only public records can be queried.

refs #59