Move controller migration to skygear-server
Originally we want to keep a minimal version of migration for gateway.
For all the changes that touch the tables that read by gateway, we
should add the revision to both skygear-server and controller. But I
found it is difficult to keep this, so move all the migrations to
skygear-server. And controller should reference the migrations in
server repo.