Update resolve gear path prefix fallback to known gears only
Originally we reserved path `/_{gear}` for gears.
This update change to route known path only, so only _auth and _asset will be routed to auth and asset gear corresponding.
For the other path that start with `/_`, we will check the deployment route.
We need this update for controller, so that we can add deployment route for controller.