Prevent infinite recursion when deriving identity verification status
Suppose OAuth identity is make verifiable, the verification status of
login ID & OAuth identity with matching email would depends on each
other. If any of them is verified, the dependency can be resolved
correctly. If all of them is not verified, it will cause infinite
recursion.

To prevent this, the dependent identity is skipped when checking for
matching email, if it has been visited before.