Prevent the same password being verified twice
The bug is due to the way we process graph input.
We iteratively feed the same input to the graph until
it reports input required.

In case of password + additional password,
the same password is being verified twice.
Very likely the second verification will fail because
the additional password should never equal to the first password.

So we need to tell the password is intended for which stage.

ref #878