Fix unexpected error when aborting signup in before_session_create hook
In hook provider, DispatchEvent will be called multiple times. If the first DispatchEvent
is success, dbHook is registered already even the latter events have error. If there is
error, all the later operations should be skipped, so we need check and skip them.
We cannot see this problem when aborting signup in before_user_create, as
before_user_create is the first event in signup, so dbHook is not be registered in this case
and no further DispatchEvent will be called.