Separate read and write audit db
In admin api, we open audit db transaction in graphql api handler for
resolve functions to query audit logs. But there are some cases that
we need to open write transaction for adding audit log (e.g. use admin
api to create user...), so we separate the stores and use different
transactions for this case.