Allow secret item to be removed
Before this change, omitting a secret item means not to update it.
But for SMTP secret, omitting it means remove it.
Therefore, we now require the whole secret config to be passed to updateApp.
If a secret item is absent, it will be deleted.
If a secret item is present but masked, it will NOT be changed.
If a secret item is present and unmasked, it will be updated.