Avoid superfluous write when panic
We wrap http.Handler inside withTx.
It could happen that the http.Handler finish its job without any error,
but the transaction cannot be committed.
In that case, a panic will occur.
The panic middleware logs the panic AND attempts to write a response again.

Some http.Handler write Content-Length explicitly.
When the panic middleware writes a response again,
ErrContentLength will occur.

Wrapping http.ResponseWriter is non-trivial because it is not the sole
interface. httpsnoop is a module specific for this non-trivial task.