Fix incorrect date range in analytics page
The bug occurs only in countries that have daylight saving.
When we use setFullYear or setDate to the date object, it applies
changes to the local time and the 1 hour shift will be automatically
adjusted. When we convert the time back to UTC, the date will become
incorrect. To avoid this problem, we always create a new time instance
in UTC when doing the date calculation.