Store bot protection verification result in authflow session
Note that current storage implementation does not support intra-flow session state mutation.
That is, if you perform more than 1 steps in the same input, steps after bot-protection-verification would not be aware of previous state changes.

We decided not to implement intra-flow session state mutation because it requires significant rework on accept and context injection logic.

We plan to implement MilestoneDidPerformBotProtectionVerification to work around this

# Conflicts:
#	pkg/lib/authenticationflow/errors.go