Set default g-recaptcha-response field as empty
Why we need this? Why not just submit this value anyway?

It's because of nginx response header max size of 4kb again.
Note currently, w/o this field, response header is around 3.5kb
If
- this field is submitted (~500-700 bytes)
- some error took place (e.g. CaptchaVerificationFailed, DuplicateLoginID)
The error, which contains the entire form value, would be returned as a cookie header.
The header will exceed maximum of 4kb, throwing 502 GateError

See removeDefaultGRecaptchaField jsdoc for more info