Fix `make venv` does not actually activate venv
venv/bin/activate does not have effect in a script.
This can be proved by adding ~/.config/pip/pip.conf

[global]
require-virtualenv = true

and then run `make venv` again.
You will see the error message of not in a virtual environment.

The simplest way to use venv in a script is to invoke the binary
directly, as done in this commit.