Name the build stages of authgear and authgear-portal uniquely
They have to be unique because we are going to use a naive file concat
to include Dockerfiles in another Dockerfile.

BuildKit supports the `# syntax=` directive to use an alternative
frontend. However, there is no good well-known and maintained frontend
that supports including Dockerfile.

So we will use the naive file concat approach.