Update .vettedpositions
---
I want you to update @./.vettedpositions for me.

This file is a list of vetted positions to mark a specific position is expected to have an lint error.

We now have an output file @./lint.output generated from the linter.

I want you to update @./.vettedpositions with the information from @./lint.output such that all lint errors are gone.